+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Proper Bandersnatch and Jubjub elliptic curves parameters. #809
 - `AdminChanged` event parameters no longer indexed. #794
+- Fix `ruint::Uint` padding for conversion to `Uint`. #808
 - Improved panic description for twisted edwards point conversions. #816
 - Fix edge case with U64 -> u128 conversion. #815
 - Conditional compilation in storage slot utilities causing build failures across different target architectures and feature combinations. #823